requirements.txt
pip
apache-airflow==2.10.2
delta-spark==3.3.2
apache-airflow-providers-apache-spark==4.11.0
requests==2.32.5
pyarrow==22.0.0
pandas==2.3.3
polars==1.34.0
glob2==0.7